refactor(ci): restore size guard headroom
