+++ crates/prodex-quota/src/render/time.rs
@@ -1,0 +1,43 @@
+use super::*;
+
+pub fn window_label(seconds: Option<i64>) -> String {
+    let Some(seconds) = seconds else {
+        return "usage".to_string();
+    };
+
+    if (17_700..=18_300).contains(&seconds) {
+        return "5h".to_string();
+    }
+    if (601_200..=608_400).contains(&seconds) {
+        return "weekly".to_string();
+    }
+    if (2_505_600..=2_678_400).contains(&seconds) {
+        return "monthly".to_string();
+    }
+
+    format!("{seconds}s")
+}
+
+pub fn format_reset_time(epoch: Option<i64>) -> String {
+    format_local_epoch(epoch, "%Y-%m-%d %H:%M")
+}
+
+pub fn format_precise_reset_time(epoch: Option<i64>) -> String {
+    format_local_epoch(epoch, "%Y-%m-%d %H:%M:%S")
+}
+
+pub fn format_quota_snapshot_time(epoch: Option<i64>) -> String {
+    format_local_epoch(epoch, "%Y-%m-%d %H:%M:%S")
+}
+
+fn format_local_epoch(epoch: Option<i64>, pattern: &str) -> String {
+    let Some(epoch) = epoch else {
+        return "-".to_string();
+    };
+
+    Local
+        .timestamp_opt(epoch, 0)
+        .single()
+        .map(|dt| dt.format(pattern).to_string())
+        .unwrap_or_else(|| epoch.to_string())
+}
